@@ -2,48 +2,38 @@
   "name": "prognosports-frontend",
   "version": "2.0.0beta",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@oruga-ui/oruga": "^0.5.5",
+    "@oruga-ui/oruga-next": "^0.5.5",
     "@oruga-ui/theme-bulma": "^0.2.5",
-    "apexcharts": "^3.35.0",
+    "@vitejs/plugin-vue": "^2.3.3",
+    "@vue-leaflet/vue-leaflet": "^0.6.1",
+    "apexcharts": "^3.35.3",
     "axios": "^0.26.1",
     "bulma-block-list": "^0.6.0",
-    "bulma-css-vars": "^0.8.0",
     "dayjs": "^1.11.0",
     "leaflet": "^1.7.1",
     "marked": "^4.0.12",
-    "sortablejs": "^1.15.0",
-    "vue": "^2.6.14",
-    "vue-apexcharts": "^1.6.2",
-    "vue-class-component": "^7.2.6",
-    "vue-clipboard2": "^0.3.3",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.5.2",
-    "vue2-leaflet": "^2.7.1",
-    "vuedraggable": "^2.24.3",
-    "vuex": "^3.6.2",
-    "vuex-class": "^0.3.2"
+    "mitt": "^3.0.0",
+    "pinia": "^2.0.14",
+    "vue": "3.2.37",
+    "vue-router": "4",
+    "vue-slicksort": "^2.0.0-alpha.5",
+    "vue3-apexcharts": "^1.4.1"
+  },
+  "resolutions": {
+    "vue-demi": "0.13.1"
   },
   "devDependencies": {
     "@types/leaflet": "^1.7.5",
     "@types/marked": "^4.0.3",
-    "@vue/cli-plugin-router": "~4.5.15",
-    "@vue/cli-plugin-typescript": "^4.5.15",
-    "@vue/cli-plugin-vuex": "^4.5.15",
-    "@vue/cli-service": "^4.5.15",
-    "node-sass": "^6.0.1",
-    "sass-loader": "^10.2.1",
+    "@types/node": "^17.0.35",
+    "sass": "^1.52.1",
     "typescript": "^4.6.3",
-    "vue-template-compiler": "^2.6.14",
-    "vuex-module-decorators": "^1.0.1"
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    "vite": "^2.9.9"
   },
   "browserslist": [
     "> 1%",
