@@ -1,14 +1,10 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-
-Vue.use(VueRouter);
+import {createRouter, createWebHistory} from 'vue-router'
 
 import {routes} from "@/_router/routes";
 import {auth, community, home, data} from "@/_router/middleware";
 
-const router = new VueRouter({
-    mode: 'history',
-    base: process.env.BASE_URL,
+const router = createRouter({
+    history: createWebHistory(import.meta.env.BASE_URL),
     routes
 });
 
