@@ -1,17 +1,21 @@
-import {NavigationGuardNext, Route} from "vue-router";
-import store from "@/_store";
+import {NavigationGuardNext, RouteLocationNormalized} from "vue-router";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
 
-export default function fetchDataFromClient(to: Route, from: Route, next: NavigationGuardNext) {
-    store.dispatch('Auth/userRequest').finally(() => {
-        const loggedIn = store.getters['Auth/isLoggedIn'];
+export default function fetchDataFromClient(to: RouteLocationNormalized, from: RouteLocationNormalized, next: NavigationGuardNext) {
+    const authStore = useAuthStore();
+    const communityStore = useCommunityStore();
+    authStore.userRequest().finally(() => {
+        const loggedIn = authStore.isLoggedIn;
+        //console.log("[🍍] Logged in? " + loggedIn)
         if (loggedIn) {
             // Si existe usuario iniciado sesión, pido su comunidad
-            const hasCommunity = store.getters['Auth/thereIsCurrentCommunity'];
-            const user = store.getters['Auth/loggedUser'];
+            const hasCommunity = communityStore.thereIsCurrentCommunity;
+            const user = authStore.loggedUser;
             if (!hasCommunity) {
-                const storedCommunityId = store.getters['Auth/storedCommunityId'];
+                const storedCommunityId = communityStore.storedCommunityId;
                 let checkForId = storedCommunityId ?? user.currentCommunity?.id;
-                store.dispatch('Auth/communityRequest', {communityId: checkForId}).finally(() => {
+                communityStore.communityRequest({communityId: checkForId}).finally(() => {
                     // Tenga o no comunidad, se ha pedido, next y se encarga otro middleware
                     next();
                 })
