@@ -1,9 +1,11 @@
-import {NavigationGuardNext, Route} from "vue-router";
-import store from "@/_store";
+import {NavigationGuardNext, RouteLocationNormalized} from "vue-router";
+import {useAuthStore} from "@/store/authStore";
 
-export default function sendToHome(to: Route, from: Route, next: NavigationGuardNext) {
+export default function sendToHome(to: RouteLocationNormalized, from: RouteLocationNormalized, next: NavigationGuardNext) {
+    const authStore = useAuthStore();
+
     // Si está iniciado sesión y entra a /, redirigir a /home
-    const loggedIn = store.getters['Auth/isLoggedIn'];
+    const loggedIn = authStore.isLoggedIn;
     if (to.path === "/" && loggedIn) {
         next('/home');
     } else {
