@@ -68,43 +68,37 @@
 </template>
 
 <script lang="ts">
-import {Component, Vue} from "vue-property-decorator";
 import PrognoPageTitle from "@/components/lib/PrognoPageTitle.vue";
-import {User} from "@/types/User";
-import {namespace} from "vuex-class";
 import AlertNoPermission from "@/components/lib/AlertNoPermission.vue";
 import {competitionService} from "@/_services";
 import {Competition} from "@/types/Competition";
 import {Season} from "@/types/Season";
 import {marked} from "marked";
-const Auth = namespace('Auth')
 
-@Component({
-        components: {
-            AlertNoPermission,
-            PrognoPageTitle,
+import {defineComponent} from "vue";
+import {useAuthStore} from "@/store/authStore";
+
+export default defineComponent({
+    name: "CompetitionEdit",
+    components: {
+        AlertNoPermission,
+        PrognoPageTitle,
+    },
+    setup() {
+        const authStore = useAuthStore();
+
+        const currentUser = authStore.user;
+        return { currentUser };
+    },
+    data() {
+        return {
+            isLoading: true,
+            thereIsCompetition: false,
+            competition: {} as Competition,
+            competitionId: this.$route.params.competition,
+            compSeasons: new Array<Season>(),
         }
-    }
-)
-export default class CompetitionEdit extends Vue {
-    @Auth.State("user") private currentUser!: User;
-
-    private isLoading = true;
-    private thereIsCompetition = false;
-    private competition?: Competition;
-    private competitionId = this.$route.params.competition;
-
-    private compSeasons: Array<Season> = [];
-
-    private isDataOk(): boolean {
-        return (this.competition !== undefined)
-            && !(this.competition.id == undefined
-                && this.competition.code == undefined
-                && this.competition.name == undefined
-                && this.competition.rules == undefined
-                && this.competition.fullname == undefined)
-    }
-
+    },
     mounted() {
         competitionService.getCompetition(this.competitionId).then((competition) => {
             this.competition = competition;
@@ -122,36 +116,47 @@ export default class CompetitionEdit extends Vue {
         }).finally(() => {
             this.isLoading = false;
         })
-    }
+    },
+    methods: {
+        isDataOk(): boolean {
+            return (this.competition !== undefined)
+                && !(this.competition.id == undefined
+                    && this.competition.code == undefined
+                    && this.competition.name == undefined
+                    && this.competition.rules == undefined
+                    && this.competition.fullname == undefined)
+        },
+        editCompetition(): void {
+            let data = {
+                code: this.competition!.code,
+                name: this.competition!.name,
+                fullname: this.competition!.fullname,
+                rules: this.competition!.rules,
+                currentSeason: this.competition!.currentSeason!.id,
+            }
+            competitionService.editCompetition(this.competition!, data).then((result) => {
+                this.$oruga.notification.open({
+                    position: 'top',
+                    message: "Se ha editado correctamente la competición `" + result.name + "`",
+                    variant: "success",
+                });
 
-    get previewRules() {
-        return marked(this.competition?.rules ?? "");
-    }
-
-    private editCompetition(): void {
-        let data = {
-            code: this.competition!.code,
-            name: this.competition!.name,
-            fullname: this.competition!.fullname,
-            rules: this.competition!.rules,
-            currentSeason: this.competition!.currentSeason!.id,
+                this.$router.push({
+                    name: 'adminCompetitions'
+                })
+            }).catch((error) => {
+                this.$oruga.notification.open({
+                    position: 'top',
+                    message: error.message,
+                    variant: "danger",
+                });
+            });
         }
-        competitionService.editCompetition(this.competition!, data).then((result) => {
-            this.$oruga.notification.open({
-                message: "Se ha editado correctamente la competición `" + result.name + "`",
-                variant: "success",
-            });
-
-            this.$router.push({
-                name: 'adminCompetitions'
-            })
-        }).catch((error) => {
-            this.$oruga.notification.open({
-                message: error.message,
-                variant: "danger",
-            });
-        });
+    },
+    computed: {
+        previewRules() {
+            return marked(this.competition?.rules ?? "");
+        }
     }
-
-}
+});
 </script>
