@@ -25,57 +25,61 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "vue-property-decorator";
 import ViewCircuitItem from "@/components/circuits/ViewCircuitItem.vue";
 import PrognoPageTitle from "@/components/lib/PrognoPageTitle.vue";
 import { Circuit } from "@/types/Circuit";
 import { circuitService } from "@/_services";
+import {defineComponent} from "vue";
 
-@Component<ViewCircuitList>({
+export default defineComponent({
+    name: "ViewCircuitList",
     components: {
         ViewCircuitItem,
         PrognoPageTitle,
     },
-})
-export default class ViewCircuitList extends Vue {
-    private circuitList: Array<Circuit> = [];
-    private filtroCircuit: String = '';
-    private isLoading: boolean = true;
-
+    data() {
+        return {
+            circuitList: new Array<Circuit>(),
+            filtroCircuit: '',
+            isLoading: true,
+        }
+    },
     created() {
         this.loadCircuits();
-    }
-
-    loadCircuits() {
-        circuitService.getCircuitList().then((circuits) => {
-            this.circuitList.push(...circuits);
-            this.isLoading = false;
-        });
-    }
-
-    get filteredCircuits(): Array<Circuit> {
-        if (!this.filtroCircuit.trim()) {
-            return this.circuitList;
+    },
+    methods: {
+        loadCircuits() {
+            circuitService.getCircuitList().then((circuits) => {
+                this.circuitList.push(...circuits);
+                this.isLoading = false;
+            });
         }
+    },
+    computed: {
+        filteredCircuits(): Array<Circuit> {
+            if (!this.filtroCircuit.trim()) {
+                return this.circuitList;
+            }
 
-        const filtroLowerCase: string = this.filtroCircuit.toLowerCase().trim();
+            const filtroLowerCase: string = this.filtroCircuit.toLowerCase().trim();
 
-        return this.circuitList.filter((circuit) => {
-            return (
-                circuit.name
-                    .toLowerCase()
-                    .includes(filtroLowerCase) ||
-                circuit.country
-                    .toLowerCase()
-                    .includes(filtroLowerCase) ||
-                circuit.locality
-                    .toLowerCase()
-                    .includes(filtroLowerCase) ||
-                circuit.variant.name
-                    .toLowerCase()
-                    .includes(filtroLowerCase)
-            );
-        });
+            return this.circuitList.filter((circuit) => {
+                return (
+                    circuit.name
+                        .toLowerCase()
+                        .includes(filtroLowerCase) ||
+                    circuit.country
+                        .toLowerCase()
+                        .includes(filtroLowerCase) ||
+                    circuit.locality
+                        .toLowerCase()
+                        .includes(filtroLowerCase) ||
+                    circuit.variant.name
+                        .toLowerCase()
+                        .includes(filtroLowerCase)
+                );
+            });
+        }
     }
-}
+});
 </script>
