@@ -14,26 +14,32 @@
 
 
 <script lang="ts">
-import { Component, Vue } from "vue-property-decorator";
 import PrognoPageTitle from "@/components/lib/PrognoPageTitle.vue";
 import {communityService} from "@/_services";
 
 import {Community} from "@/types/Community";
-import {namespace} from "vuex-class";
-const Auth = namespace('Auth')
 
-@Component<ViewOneCommunity>({
+import {defineComponent} from "vue";
+import {useCommunityStore} from "@/store/communityStore";
+
+export default defineComponent({
+    name: "ViewOneCommunity",
     components: {
         PrognoPageTitle
     },
-})
-export default class ViewOneCommunity extends Vue {
-    @Auth.Action setCommunity!: (community: Community) => void;
+    setup() {
+        const communityStore = useCommunityStore();
 
-    private community!: Community;
-    private isLoading: boolean = true;
-    private thereIsCommunity: boolean = false;
-
+        const setCommunity = communityStore.setCommunity;
+        return { setCommunity };
+    },
+    data() {
+        return {
+            community: {} as Community,
+            isLoading: true,
+            thereIsCommunity: false,
+        }
+    },
     created() {
         let communityId = this.$route.params.community;
         let code = this.$route.params.code;
@@ -45,6 +51,7 @@ export default class ViewOneCommunity extends Vue {
             communityService.joinCommunity(community, code).then((communityRes) => {
                 setTimeout(() => {
                     this.$oruga.notification.open({
+                        position: 'top',
                         message: "¡Te has unido correctamente a " + communityRes.name + "!",
                         variant: "success",
                     });
@@ -59,6 +66,7 @@ export default class ViewOneCommunity extends Vue {
             }).catch((reason => {
                 setTimeout(() => {
                     this.$oruga.notification.open({
+                        position: 'top',
                         message: "Ha ocurrido un error: " + reason.message,
                         variant: "danger",
                     });
@@ -72,17 +80,15 @@ export default class ViewOneCommunity extends Vue {
         }).finally(() => {
             this.isLoading = false;
         })
-    }
-
-    get communityName() {
-        if (this.thereIsCommunity) {
-            return this.community.name;
-        } else {
-            return "Comunidad no encontrada";
+    },
+    computed: {
+        communityName() {
+            if (this.thereIsCommunity) {
+                return this.community.name;
+            } else {
+                return "Comunidad no encontrada";
+            }
         }
     }
-}
+});
 </script>
-
-<style scoped>
-</style>
