@@ -34,74 +34,73 @@
                     </o-field>
                 </div>
 
-                <draggable class="block-list has-radius is-highlighted is-info no-select"
-                           :list="pilotosDisponiblesFiltrados" :group="session.name">
-                    <transition-group type="transition" :name="!drag ? 'flip-list' : null">
-                        <li
-                            class="is-highlighted has-text-weight-semibold is-flex is-justify-content-space-between"
-                            :style="styleDriverCard(element)"
-                            v-for="(element, index) in pilotosDisponiblesFiltrados"
-                            :key="element.number">
+                <SlickList v-model:list="pilotosDisponibles" :group="session.name" :accept="[session.name]" tag="ul" :distance="1"
+                           class="block-list has-radius no-select">
 
-                            <span>
-                                {{ element.firstname }} {{ element.lastname }}
-                                <span class="tag is-rounded" v-bind:style="styleDorsal(element)">#{{ element.number }}</span>
-                                <o-tooltip class="ml-1" :label="element.team.longname">
-                                    {{ element.team.name }}
-                                </o-tooltip>
-                                ({{element.team.carname}})
-                            </span>
+                    <SlickItem v-for="(item, index) in pilotosDisponiblesFiltrados" :key="item.id" :index="index" tag="li"
+                               class="is-highlighted has-text-weight-semibold is-flex is-justify-content-space-between"
+                               :style="styleDriverCard(item)">
+                        <span>
+                            {{ item.firstname }} {{ item.lastname }}
+                            <span class="tag is-rounded" v-bind:style="styleDorsal(item)">#{{ item.number }}</span>
+                            <o-tooltip class="ml-1" :label="item.team.longname">
+                                {{ item.team.name }}
+                            </o-tooltip>
+                            ({{item.team.carname}})
+                        </span>
 
-                            <a @click="moveToTippList(element, index)" class="pl-3 pr-3">
-                                <i class="mr-0 fas fa-angle-right has-text-primary"></i>
-                                <i class="ml-0 fas fa-angle-right has-text-primary"></i>
-                            </a>
-                        </li>
-                    </transition-group>
-                </draggable>
+                        <a @click="moveToTippList(item, index)" class="pl-3 pr-3 arrow-col has-text-primary">
+                            <i class="is-hidden-touch mr-0 fas fa-angle-right"></i>
+                            <i class="is-hidden-touch ml-0 fas fa-angle-right"></i>
+                            <i class="is-hidden-desktop mt-0 fas fa-angle-up"></i>
+                            <i class="is-hidden-desktop mb-0 fas fa-angle-up"></i>
+                        </a>
+                    </SlickItem>
+                </SlickList>
+
             </div>
+
             <div class="column is-6">
-                <h3>Pilotos pronosticados ({{ cantidadPilotosPronosticados(ruleSet, session) }})</h3>
-                <draggable class="block-list has-radius is-highlighted is-primary no-select"
-                           :list="pilotosPronosticados" :group="session.name"
-                           :emptyInsertThreshold="1000">
-                    <transition-group type="transition" tag="div" :name="!drag ? 'flip-list' : null">
-                        <li
-                            class="is-highlighted has-text-weight-semibold is-flex is-justify-content-left"
-                            v-bind:style="styleDriverCard(element)"
-                            v-for="(element, index)  in pilotosPronosticados"
-                            :key="element.number">
+                <h3>Pilotos pronosticados</h3>
+                <SlickList v-model:list="pilotosPronosticados" :group="session.name" :accept="[session.name]" tag="ul" :distance="1"
+                           class="block-list no-select">
 
-                            <a @click="moveToAvailableList(element, index)" class="mr-3 pl-3">
-                                <i class="mr-0 fas fa-angle-left has-text-primary"></i>
-                                <i class="ml-0 fas fa-angle-left has-text-primary"></i>
-                            </a>
+                    <SlickItem v-for="(item, index) in pilotosPronosticados" :key="item.id" :index="index" tag="li"
+                               class="is-highlighted has-text-weight-semibold has-radius is-flex is-justify-content-left"
+                               :style="styleDriverCard(item)">
 
-                            <span>
-                                <b>{{ index + 1 }}º.</b> {{ element.firstname }} {{ element.lastname }}
-                                <span class="tag is-rounded" v-bind:style="styleDorsal(element)">#{{ element.number }}</span>
-                                <o-tooltip class="ml-1" :label="element.team.longname">
-                                    {{ element.team.name }}
-                                </o-tooltip>
-                                ({{element.team.carname}})
-                            </span>
-                        </li>
-                    </transition-group>
-                </draggable>
+                        <a @click="moveToAvailableList(item, index)" class="mr-3 pl-3 arrow-col has-text-primary">
+                            <i class="is-hidden-touch mr-0 fas fa-angle-left"></i>
+                            <i class="is-hidden-touch ml-0 fas fa-angle-left"></i>
+                            <i class="is-hidden-desktop mt-0 fas fa-angle-down"></i>
+                            <i class="is-hidden-desktop mb-0 fas fa-angle-down"></i>
+                        </a>
+
+                        <span>
+                            <b>{{ index + 1 }}º.</b> {{ item.firstname }} {{ item.lastname }}
+                            <span class="tag is-rounded" v-bind:style="styleDorsal(item)">#{{ item.number }}</span>
+                            <o-tooltip class="ml-1" :label="item.team.longname">
+                                {{ item.team.name }}
+                            </o-tooltip>
+                            ({{item.team.carname}})
+                        </span>
+                    </SlickItem>
+                </SlickList>
+
                 <template v-if="isBeforeEndDate(this.session)">
                     <o-button v-if="pilotosPronosticados.length === cantidadPilotosPronosticados(ruleSet, session)"
                               variant="success is-fullwidth"
                               @click="enviarPronostico">Enviar pronóstico
                     </o-button>
+
+                    <div v-else class="notification is-warning is-light">
+                        El pronóstico debe tener {{ cantidadPilotosPronosticados(ruleSet, session) }} pilotos escogidos y ordenados.
+                    </div>
                 </template>
+
                 <o-button v-else disabled variant="success is-fullwidth">
                     Ya no se puede pronosticar
                 </o-button>
-
-
-                <div v-else class="notification is-warning is-light">
-                    El pronóstico debe tener {{ cantidadPilotosPronosticados(ruleSet, session) }} pilotos escogidos y ordenados.
-                </div>
 
                 <hr v-if="(pilotosPronosticados.length > 0) && isBeforeEndDate(this.session)"/>
                 <o-button v-if="(pilotosPronosticados.length > 0) && isBeforeEndDate(this.session)" variant="danger is-light is-fullwidth" @click="reiniciarPronostico">Limpiar pronóstico</o-button>
@@ -112,46 +111,68 @@
 </template>
 
 <script lang="ts">
-import {Component, Prop, Vue} from "vue-property-decorator";
 import {RaceSession} from "@/types/RaceSession";
-import draggable from "vuedraggable";
+import { SlickList, SlickItem } from 'vue-slicksort';
 import {grandPrixService} from "@/_services";
 import {GrandPrix} from "@/types/GrandPrix";
 import {Driver} from "@/types/Driver";
-import {Community} from "@/types/Community";
-import {namespace} from "vuex-class";
 import {RaceResult} from "@/types/RaceResult";
-import {User} from "@/types/User";
 import {StartGridPosition} from "@/types/StartGridPosition";
 import {RuleSet} from "@/types/RuleSet";
-const Auth = namespace('Auth')
 
-@Component({
+import {defineComponent, PropType} from "vue";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
+
+export default defineComponent({
+    name: "SelectTipps",
     components: {
-        draggable
-    }}
-)
-export default class SelectTipps extends Vue {
-    @Prop({required: true}) session!: RaceSession;
-    @Prop({required: true}) grandPrix!: GrandPrix;
-    @Prop({required: true}) ruleSet!: RuleSet;
-    @Prop({required: true}) drivers!: Array<Driver>;
-    @Prop({required: true}) startGrids!: Map<RaceSession, Array<StartGridPosition>>;
+        SlickList,
+        SlickItem,
+    },
+    props: {
+        session: {
+            type: Object as PropType<RaceSession>,
+            required: true,
+        },
+        grandPrix: {
+            type: Object as PropType<GrandPrix>,
+            required: true,
+        },
+        ruleSet: {
+            type: Object as PropType<RuleSet>,
+            required: true,
+        },
+        drivers: {
+            type: Array as PropType<Array<Driver>>,
+            required: true,
+        },
+        startGrids: {
+            type: Map as PropType<Map<RaceSession, Array<StartGridPosition>>>,
+            required: true,
+        },
+    },
+    setup() {
+        const authStore = useAuthStore();
+        const communityStore = useCommunityStore();
 
-    @Auth.State("user") private currentUser!: User;
-    @Auth.State("community") private currentCommunity!: Community;
+        const currentUser = authStore.user;
+        const currentCommunity = communityStore.community;
+        return { currentUser, currentCommunity };
+    },
+    data() {
+        return {
+            drag: false,
+            filtroPiloto: '',
+            orderType: 1,
+            orderAscendent: false,
 
-    private drag: boolean = false;
-    private filtroPiloto: string = '';
-    private orderType: number = 1;
-    private orderAscendent: boolean = false;
-
-    private pilotosPronosticados: Array<Driver> = [];
-    private pilotosDisponibles: Array<Driver> = [];
-    private originalPilotos: Array<Driver> = [];
-
-    private indexedGrid: Map<number, number> = new Map(); // Dorsal del piloto -> Posicion en la grid
-
+            pilotosPronosticados: new Array<Driver>(),
+            pilotosDisponibles: new Array<Driver>(),
+            originalPilotos: new Array<Driver>(),
+            indexedGrid: new Map<number, number>(), // Dorsal del piloto -> Posicion en la grid
+        }
+    },
     mounted() {
 
         this.originalPilotos.push(...this.drivers);
@@ -179,106 +200,105 @@ export default class SelectTipps extends Vue {
                 }
             }
         });
-    }
+    },
+    methods: {
+        moveToTippList(driver: Driver, index: number) {
+            console.log("Moviendo a pronosticados " + driver.code);
+            this.pilotosDisponibles.splice(index, 1);
+            this.pilotosPronosticados.push(driver);
+        },
+        moveToAvailableList(driver: Driver, index: number) {
+            console.log("Moviendo a disponibles " + driver.code);
+            this.pilotosPronosticados.splice(index, 1);
+            this.pilotosDisponibles.push(driver);
+        },
+        styleDriverCard(driver: Driver) {
+            return {
+                color: 'black',
+                'border': '1px solid #'+ driver.team.teamcolor,
+                'border-left': '10px #'+ driver.team.teamcolor + ' solid',
+                // 'border-right': '30px #'+ driver.team.teamcolor + ' solid',
+                'border-right-image-source': 'linear-gradient(to left, #'+ driver.team.teamcolor + ', #ffffff)',
+                opacity: 0.9,
+            }
+        },
+        styleDorsal(driver: Driver) {
+            return {
+                color: 'white',
+                textShadow: '0 0 2px #000',
+                backgroundColor: '#' + driver.team.teamcolor,
+            }
+        },
+        reiniciarPronostico() {
+            this.pilotosPronosticados = [];
+            this.pilotosDisponibles = [];
+            this.pilotosDisponibles.push(...this.originalPilotos);
+        },
+        enviarPronostico() {
+            let tipps: Array<RaceResult> = [];
 
-    public moveToTippList(driver: Driver, index: number) {
-        console.log("Moviendo a pronosticados " + driver.code);
-        this.pilotosDisponibles.splice(index, 1);
-        this.pilotosPronosticados.push(driver);
-    }
-
-    public moveToAvailableList(driver: Driver, index: number) {
-        console.log("Moviendo a disponibles " + driver.code);
-        this.pilotosPronosticados.splice(index, 1);
-        this.pilotosDisponibles.push(driver);
-    }
-
-    public styleDriverCard(driver: Driver) {
-        return {
-            color: 'black',
-            'border': '1px solid #'+ driver.team.teamcolor,
-            'border-left': '10px #'+ driver.team.teamcolor + ' solid',
-            // 'border-right': '30px #'+ driver.team.teamcolor + ' solid',
-            'border-right-image-source': 'linear-gradient(to left, #'+ driver.team.teamcolor + ', #ffffff)',
-            opacity: 0.9,
-        }
-    }
-
-    public styleDorsal(driver: Driver) {
-        return {
-            color: 'white',
-            textShadow: '0 0 2px #000',
-            backgroundColor: '#' + driver.team.teamcolor,
-        }
-    }
-
-    get pilotosDisponiblesFiltrados(): Array<Driver> {
-        const sortAlfabetico = (d1: Driver, d2: Driver) => (d1.lastname < d2.lastname ? -1 : 1);
-        const sortEquipos = (d1: Driver, d2: Driver) => (d1.team.name < d2.team.name ? -1 : 1);
-        const sortDorsal = (d1: Driver, d2: Driver) => (d1.number < d2.number ? -1 : 1);
-        const sortParrilla = (d1: Driver, d2: Driver) => (this.indexedGrid.get(d1.number)! < this.indexedGrid.get(d2.number)! ? -1 : 1);
-
-
-        let pickedSort: (d1: Driver, d2: Driver) => (number);
-        switch (this.orderType) {
-            case 1: pickedSort = sortEquipos; break;
-            case 2: pickedSort = sortDorsal; break;
-            case 3: pickedSort = sortParrilla; break;
-            default: pickedSort = sortAlfabetico;
-        }
-        let listaOrdenada = this.pilotosDisponibles.sort(pickedSort);
-
-        if (this.orderAscendent) {
-            listaOrdenada = listaOrdenada.reverse();
-        }
-        if (!this.filtroPiloto.trim()) {
-            return listaOrdenada;
-        }
-
-        let filtroLowerCase = this.filtroPiloto.toLowerCase();
-        return listaOrdenada.filter(driver => {
-            return driver.lastname.toLowerCase().includes(filtroLowerCase)
-                || driver.firstname.toLowerCase().includes(filtroLowerCase)
-                || driver.team.name.toLowerCase().includes(filtroLowerCase)
-                || driver.team.longname.toLowerCase().includes(filtroLowerCase)
-                || driver.team.carname.toLowerCase().includes(filtroLowerCase);
-        })
-    }
-
-    private reiniciarPronostico() {
-        this.pilotosPronosticados = [];
-        this.pilotosDisponibles = [];
-        this.pilotosDisponibles.push(...this.originalPilotos);
-    }
-
-    private enviarPronostico() {
-        let tipps: Array<RaceResult> = [];
-
-        this.pilotosPronosticados.forEach((driver, index, array) => {
-            tipps.push({position: index + 1, driver: {id: driver.id} as Driver} as RaceResult);
-        });
-        grandPrixService.postUserTipps(this.grandPrix, this.session, this.currentCommunity, tipps).then(
-            () => {
-                this.$oruga.notification.open({
-                    message: "¡Has guardado tus pronósticos!",
-                    variant: "success",
-                });
-            },
-            (error: any) => {
-                let message = "Error guardando tus pronósticos: " + error.message;
-
-                this.$oruga.notification.open({
-                    duration: 5000,
-                    message: message,
-                    variant: "danger",
-                });
+            this.pilotosPronosticados.forEach((driver, index, array) => {
+                tipps.push({position: index + 1, driver: {id: driver.id} as Driver} as RaceResult);
             });
+            grandPrixService.postUserTipps(this.grandPrix, this.session, this.currentCommunity, tipps).then(
+                () => {
+                    this.$oruga.notification.open({
+                        position: 'top',
+                        message: "¡Has guardado tus pronósticos!",
+                        variant: "success",
+                    });
+                },
+                (error: any) => {
+                    let message = "Error guardando tus pronósticos: " + error.message;
+
+                    this.$oruga.notification.open({
+                        position: 'top',
+                        duration: 5000,
+                        message: message,
+                        variant: "danger",
+                    });
+                });
+        }
+    },
+    computed: {
+        pilotosDisponiblesFiltrados(): Array<Driver> {
+            const sortAlfabetico = (d1: Driver, d2: Driver) => (d1.lastname < d2.lastname ? -1 : 1);
+            const sortEquipos = (d1: Driver, d2: Driver) => (d1.team.name < d2.team.name ? -1 : 1);
+            const sortDorsal = (d1: Driver, d2: Driver) => (d1.number < d2.number ? -1 : 1);
+            const sortParrilla = (d1: Driver, d2: Driver) => (this.indexedGrid.get(d1.number)! < this.indexedGrid.get(d2.number)! ? -1 : 1);
+
+
+            let pickedSort: (d1: Driver, d2: Driver) => (number);
+            switch (this.orderType) {
+                case 1: pickedSort = sortEquipos; break;
+                case 2: pickedSort = sortDorsal; break;
+                case 3: pickedSort = sortParrilla; break;
+                default: pickedSort = sortAlfabetico;
+            }
+            let listaOrdenada = this.pilotosDisponibles.sort(pickedSort);
+
+            if (this.orderAscendent) {
+                listaOrdenada = listaOrdenada.reverse();
+            }
+            if (!this.filtroPiloto.trim()) {
+                return listaOrdenada;
+            }
+
+            let filtroLowerCase = this.filtroPiloto.toLowerCase();
+            return listaOrdenada.filter(driver => {
+                return driver.lastname.toLowerCase().includes(filtroLowerCase)
+                    || driver.firstname.toLowerCase().includes(filtroLowerCase)
+                    || driver.team.name.toLowerCase().includes(filtroLowerCase)
+                    || driver.team.longname.toLowerCase().includes(filtroLowerCase)
+                    || driver.team.carname.toLowerCase().includes(filtroLowerCase);
+            })
+        }
     }
-}
+});
 </script>
 
 <style scoped lang="scss">
-@import "~bulma/sass/utilities/_all";
+@import "bulma/sass/utilities/_all.sass";
 
 .block-list:empty,
 .block-list > div:empty {
@@ -302,6 +322,12 @@ export default class SelectTipps extends Vue {
     .listas {
         display: flex;
         flex-direction: column-reverse;
+    }
+}
+@media screen and (max-width: 1023px) {
+    .arrow-col {
+        display: flex;
+        flex-direction: column;
     }
 }
 </style>
