@@ -119,14 +119,20 @@
     </div>
 </template>
 
+<style lang="scss" scoped>
+@import "@/scss/landing.scss";
+</style>
+
 <script lang="ts">
-import {Component, Vue} from "vue-property-decorator";
+import {defineComponent} from "vue";
 
-@Component
-export default class LandingHome extends Vue {
-
-    sendToRegister() {
-        this.$router.push({path: '/register'});
+export default defineComponent({
+    name: "LandingHome",
+    methods: {
+        sendToRegister() {
+            window.scrollTo({ top: 0, left: 0, behavior: "smooth"});
+            this.$router.push({path: '/register'});
+        }
     }
-}
+});
 </script>
