@@ -115,115 +115,136 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue } from "vue-property-decorator";
 import {userService} from "@/_services";
 
-@Component
-export default class ForgotPasswordComponent extends Vue {
-    private email: string = "";
-    private inputToken: string = "";
-    private inputPassword: string = "";
-    private showChangePasswordForm: boolean = false;
+import {defineComponent} from "vue";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
 
-    handleSubmitChangePassword() {
-        if (this.email) {
-            userService.changePassword(
-                this.email,
-                this.inputToken,
-                this.inputPassword
-            ).then(
-                () => {
-                    this.$oruga.notification.open({
-                        message: "Tu contraseña ha sido restablecida",
-                        variant: "success",
-                    });
-                    this.$router.push({
-                        path: '/login',
-                        query: {redirect: this.$route.query.redirect}
-                    });
-                },
-                (error) => {
-                    if (error === "User email cannot be null") {
+export default defineComponent({
+    name: "LandingNavbar",
+    setup() {
+        const authStore = useAuthStore();
+        const communityStore = useCommunityStore();
+
+        const currentUser = authStore.user;
+        const currentCommunity = communityStore.community;
+        return { currentUser, currentCommunity };
+    },
+    data() {
+        return {
+            email: "",
+            inputToken: "",
+            inputPassword: "",
+            showChangePasswordForm: false
+        }
+    },
+    methods: {
+        handleSubmitChangePassword() {
+            if (this.email) {
+                userService.changePassword(
+                    this.email,
+                    this.inputToken,
+                    this.inputPassword
+                ).then(
+                    () => {
                         this.$oruga.notification.open({
-                            duration: 5000,
-                            message:
-                                "Debes introducir tu dirección de email",
-                            variant: "danger",
+                            position: 'top',
+                            message: "Tu contraseña ha sido restablecida",
+                            variant: "success",
                         });
-                    } else if (error === "User not found") {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message: "Usuario no encontrado",
-                            variant: "danger",
+                        this.$router.push({
+                            path: '/login',
+                            query: {redirect: this.$route.query.redirect}
                         });
-                    } else if (error === "You must send the security token") {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message:
-                                "Debes escribir el código de seguridad recibido",
-                            variant: "danger",
-                        });
-                    } else if (error === "You must send new the password") {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message: "Debes escribir tu nueva contraseña",
-                            variant: "danger",
-                        });
-                    } else if (error === "Token rejected") {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message:
-                                "Token rechazado. Compruebalo bien o vuelve a intentarlo en 15 minutos",
-                            variant: "danger",
-                        });
-                    } else {
-                        console.log(error)
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message: "Ha ocurrido desconocido cambiando la contraseña",
-                            variant: "danger",
-                        });
+                    },
+                    (error) => {
+                        if (error === "User email cannot be null") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Debes introducir tu dirección de email",
+                                variant: "danger",
+                            });
+                        } else if (error === "User not found") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Usuario no encontrado",
+                                variant: "danger",
+                            });
+                        } else if (error === "You must send the security token") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Debes escribir el código de seguridad recibido",
+                                variant: "danger",
+                            });
+                        } else if (error === "You must send new the password") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Debes escribir tu nueva contraseña",
+                                variant: "danger",
+                            });
+                        } else if (error === "Token rejected") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Token rechazado. Compruebalo bien o vuelve a intentarlo en 15 minutos",
+                                variant: "danger",
+                            });
+                        } else {
+                            console.log(error)
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Ha ocurrido desconocido cambiando la contraseña",
+                                variant: "danger",
+                            });
+                        }
                     }
-                }
-            );
+                );
+            }
+        },
+        handleSendCode() {
+            if (this.email) {
+                userService.sendForgotPassword(this.email).then(
+                    () => {
+                        this.$oruga.notification.open({
+                            position: 'top',
+                            message: "Tu código de verificación ha sido enviado",
+                            variant: "success",
+                        });
+                        this.showChangePasswordForm = true;
+                    },
+                    (error) => {
+                        if (error === "User email cannot be null") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Debes introducir tu dirección de email",
+                                variant: "danger",
+                            });
+                        } else if (error === "User not found") {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Usuario no encontrado",
+                                variant: "danger",
+                            });
+                        } else {
+                            this.$oruga.notification.open({
+                                position: 'top',
+                                duration: 5000,
+                                message: "Ha ocurrido un error solicitando el código",
+                                variant: "danger",
+                            });
+                        }
+                    }
+                );
+            }
         }
     }
-
-    handleSendCode() {
-        if (this.email) {
-            userService.sendForgotPassword(this.email).then(
-                () => {
-                    this.$oruga.notification.open({
-                        message:
-                            "Tu código de verificación ha sido enviado",
-                        variant: "success",
-                    });
-                    this.showChangePasswordForm = true;
-                },
-                (error) => {
-                    if (error === "User email cannot be null") {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message:
-                                "Debes introducir tu dirección de email",
-                            variant: "danger",
-                        });
-                    } else if (error === "User not found") {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message: "Usuario no encontrado",
-                            variant: "danger",
-                        });
-                    } else {
-                        this.$oruga.notification.open({
-                            duration: 5000,
-                            message: "Ha ocurrido un error solicitando el código",
-                            variant: "danger",
-                        });
-                    }
-                }
-            );
-        }
-    }
-}
+});
 </script>
