@@ -3,21 +3,27 @@
 </template>
 
 <script lang="ts">
-import {Component, Vue} from "vue-property-decorator";
-import {namespace} from "vuex-class";
-const Auth = namespace("Auth");
-@Component
-export default class LogoutComponent extends Vue {
 
-    @Auth.Action private signOut!: () => Promise<void>;
+import {useAuthStore} from "@/store/authStore";
+import {defineComponent} from "vue";
+
+export default defineComponent({
+    name: "LogoutComponent",
+    setup() {
+        const authStore = useAuthStore();
+
+        const signOut = authStore.signOut;
+        return { signOut };
+    },
     created() {
         this.signOut().then(() => {
             this.$router.push({name: 'login'});
             this.$oruga.notification.open({
+                position: 'top',
                 message: 'Has cerrado sesión correctamente',
-                type: 'is-info'
+                type: 'info'
             });
         });
     }
-};
+});
 </script>
