@@ -28,8 +28,8 @@
                             </section>
                             <footer class="modal-card-foot">
                                 <span class="has-text-info is-italic has-text-right">Recibida el
-                                    <o-tooltip :label="noti.createdAt | dateDiff" position="is-right">
-                                        {{ noti.createdAt | humanDateTime }}
+                                    <o-tooltip :label="dateDiff(noti.createdAt)" position="right">
+                                        {{ humanDateTime(noti.createdAt) }}
                                     </o-tooltip>
                                 </span>
                             </footer>
@@ -41,43 +41,61 @@
 </template>
 
 <script lang="ts">
-    import {Component, Vue} from "vue-property-decorator";
     import {notificationService} from "@/_services";
     import { Notification } from "@/types/Notification";
-    import Oruga from "@oruga-ui/oruga";
 
-    @Component
-    export default class NotificationsDropdown extends Vue {
-        private unreadNotificationsCount = 0;
-        private notificationList: Array<Notification> = [];
-        private leidas: boolean = false;
+    import {useAuthStore} from "@/store/authStore";
+    import {useCommunityStore} from "@/store/communityStore";
+    import {defineComponent} from "vue";
+    import {useDayjs} from "@/composables/useDayjs";
 
+    export default defineComponent({
+        name: "NotificationsDropdown",
+        setup() {
+            const dayjs = useDayjs();
+            const authStore = useAuthStore();
+            const communityStore = useCommunityStore();
+
+            const dateDiff = dayjs.dateDiff;
+            const humanDateTime = dayjs.humanDateTime;
+            const currentUser = authStore.user;
+            const currentCommunity = communityStore.community;
+            return { currentUser, currentCommunity, dateDiff, humanDateTime };
+        },
+        data() {
+            return {
+                unreadNotificationsCount: 0,
+                notificationList: [] as Array<Notification>,
+                leidas: false
+            }
+        },
         created() {
             this.getNotifications();
-        }
-
-        public clearNotification() {
-            Vue.prototype.$oruga.notification.open({
-                message: "Has limpiado tus notificaciones",
-                variant: "success",
-            });
-            notificationService.clearNotifications().then(() => {
-                this.unreadNotificationsCount = 0;
-                this.leidas = true;
-            });
-        }
-
-        public getNotifications()  {
-            notificationService.getNotifications().then((result) => {
-                this.notificationList = result;
-                this.notificationList.forEach(notification => {
-                    if (notification.readAt == undefined) {
-                        this.unreadNotificationsCount++;
-                    }
+        },
+        methods: {
+            clearNotification() {
+                this.$oruga.notification.open({
+                    position: 'top',
+                    message: "Has limpiado tus notificaciones",
+                    variant: "success",
                 });
-            });
+                notificationService.clearNotifications().then(() => {
+                    this.unreadNotificationsCount = 0;
+                    this.leidas = true;
+                });
+            },
+            getNotifications() {
+                notificationService.getNotifications().then((result) => {
+                    this.notificationList = result;
+                    this.notificationList.forEach(notification => {
+                        if (notification.readAt == undefined) {
+                            this.unreadNotificationsCount++;
+                        }
+                    });
+                });
+            }
         }
-    }
+    });
 </script>
 
 <style lang="css" scoped>
