@@ -33,7 +33,7 @@
                     <span class="icon is-small">
                         <i class="fas fa-clock fa-sm mr-2"></i>
                     </span>
-                    <span class="multiline-text">Visto por última vez {{ user.last_activity | dateDiff }}</span>
+                    <span class="multiline-text">Visto por última vez {{ dateDiff(user.last_activity) }}</span>
                 </p>
             </div>
         </div>
@@ -41,13 +41,24 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "vue-property-decorator";
 import { User } from "@/types/User";
+import {defineComponent, PropType} from "vue";
+import {useDayjs} from "@/composables/useDayjs";
 
-@Component
-export default class ScoreComponents extends Vue {
-    @Prop({ required: true }) user!: User;
-}
+export default defineComponent({
+    name: "ScoreComponents",
+    props: {
+        user: {
+            type: Object as PropType<User>,
+            required: true,
+        }
+    },
+    setup() {
+        const dayjs = useDayjs();
+        const dateDiff = dayjs.dateDiff;
+        return { dateDiff };
+    }
+});
 </script>
 
 <style lang="css" scoped>
