@@ -51,7 +51,7 @@
                                         <span class="icon mr-2">
                                             <i class="fas fa-clock"></i>
                                         </span>
-                                        Última conexión: {{ profile.last_activity | dateDiff }}
+                                        Última conexión: {{ dateDiff(profile.last_activity) }}
                                     </span>
                                 </div>
                                 <div class="block mb-1">
@@ -59,7 +59,7 @@
                                         <span class="icon mr-2">
                                             <i class="fas fa-calendar"></i>
                                         </span>
-                                        Registrado el {{ profile.created | humanDateTime }}
+                                        Registrado el {{ humanDateTime(profile.created) }}
                                     </span>
                                 </div>
                                 <div v-if="profile.location" class="block mb-1">
@@ -75,7 +75,7 @@
                                         <span class="icon mr-2">
                                             <i class="fas fa-birthday-cake"></i>
                                         </span>
-                                        {{ profile.birthdate | humanDate }}
+                                        {{ humanDate(profile.birthdate) }}
                                     </span>
                                 </div>
                             </div>
@@ -97,31 +97,43 @@
 </template>
 
 <script lang="ts">
-import {Component, Vue} from "vue-property-decorator";
 import {User, UserResume} from "@/types/User";
-import {namespace} from "vuex-class";
-import EventBus from "@/plugins/eventbus";
 import {userService} from "@/_services";
 import {Season} from "@/types/Season";
-import {Community} from "@/types/Community";
 import UserLevelResume from "@/components/user/UserLevelResume.vue";
 
-const Auth = namespace('Auth')
+import {defineComponent} from "vue";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
+import useEmitter from "@/composables/useEmitter";
+import {useDayjs} from "@/composables/useDayjs";
 
-@Component({
+export default defineComponent({
+    name: "UserProfile",
     components: {
         UserLevelResume
-    }
-})
-export default class UserProfile extends Vue {
-    @Auth.State("user") private currentUser!: User;
-    @Auth.State("community") private currentCommunity!: Community;
+    },
+    setup() {
+        const dayjs = useDayjs();
+        const emitter = useEmitter();
+        const authStore = useAuthStore();
+        const communityStore = useCommunityStore();
 
-    private thereIsUserParam = false;
-    private isLoading = true;
-    private profile: User | null = null;
-    private userResume: UserResume | null = null;
-
+        const dateDiff = dayjs.dateDiff;
+        const humanDateTime = dayjs.humanDateTime;
+        const humanDate = dayjs.humanDate;
+        const currentUser = authStore.user;
+        const currentCommunity = communityStore.community;
+        return {currentUser, currentCommunity, emitter, dateDiff, humanDateTime, humanDate};
+    },
+    data() {
+        return {
+            thereIsUserParam: false,
+            isLoading: true,
+            profile: {} as User | null,
+            userResume: null as UserResume | null,
+        }
+    },
     mounted() {
         const findProfile = new Promise<User | null>((resolve, reject) => {
             // Si se ha buscado un user en la URL, se procesa
@@ -129,7 +141,8 @@ export default class UserProfile extends Vue {
                 this.thereIsUserParam = true;
                 userService.getUser(this.$route.params.user).then((user) => {
                     this.profile = user;
-                }).catch(()=>{}).finally(() => {
+                }).catch(() => {
+                }).finally(() => {
                     this.isLoading = false;
 
                     resolve(this.profile);
@@ -152,14 +165,13 @@ export default class UserProfile extends Vue {
                 this.userResume = resume;
             });
         });
+    },
+    methods: {
+        changeBreadcrumb(name: string) {
+            this.emitter.emit('breadcrumbLastname', name);
+        }
     }
-
-
-    changeBreadcrumb(name: string) {
-        EventBus.$emit('breadcrumbLastname', name);
-    }
-
-}
+});
 </script>
 
 <style lang="css">
