@@ -3,10 +3,9 @@
 </template>
 
 <script lang="ts">
-import {Component, Vue} from "vue-property-decorator";
+import {defineComponent} from "vue";
 
-@Component
-export default class EmptyRoutedView extends Vue {
-
-}
+export default defineComponent({
+    name: "EmptyRouterView"
+});
 </script>
